add separate tab for image upload
