+++ src/app/trader/profile/profile.component.scss
@@ -1,12 +1,5 @@
 @import '../../../theme/variables.scss';
 
-.profile-wrapper {
-  max-width: 1024px;
-  margin: 0 auto;
-  padding: 0 16px;
-  box-sizing: border-box;
-}
-
 .verify-mail-alert {
   background-color: lightgoldenrodyellow;
   border: 1px solid lightsalmon;
@@ -15,125 +8,4 @@
   border-radius: 15px;
   margin-bottom: 20px;
   text-align: center;
-}
-
-div {
-  margin-top: 2rem;
-}
-
-.move-right {
-  margin-left: 1rem;
-}
-
-.dense {
-  margin-top: 0.4rem;
-}
-
-input {
-  min-width: 10px;
-}
-
-.warning {
-  color: $errorcolor;
-  font-size: 0.7rem;
-  margin-top: 1px;
-  text-align: center;
-}
-
-.image-list {
-  width: 100%;
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(150px, 1fr));
-  grid-gap: 1.5rem;
-  margin-bottom: 20px;
-  margin-top: 10px;
-
-  .image-item {
-    margin: 0px;
-    background-size: cover;
-    border: dotted 1px $lightergreen;
-    border-radius: 10px;
-    position: relative;
-
-    &::before {
-      content: '';
-      padding-bottom: 100%;
-      display: inline-block;
-      vertical-align: top;
-    }
-
-    .btns {
-      position: absolute;
-      bottom: 10px;
-      left: 0;
-      right: 0;
-      text-align: center;
-    }
-
-    // .delete-button {
-    //   bottom: 10px;
-    //   position: absolute;
-    //   width: 100%;
-
-    //   lk-button {
-    //     display: block;
-    //     margin: 0 auto;
-    //     width: max-content;
-    //   }
-    // }
-  }
-}
-
-.button-and-success {
-  display: flex;
-  flex-direction: column;
-  margin: 0 auto;
-}
-
-.mat-mini-fab.mat-accent {
-  background-color: #fff;
-  outline: none;
-  font-size: 0.8rem;
-  display: inline-block;
-  margin-left: 10px;
-  margin-right: 10px;
-}
-
-.success {
-  margin-top: 0px;
-  margin-left: 20px;
-  margin-bottom: 20px;
-  color: $darkgreen;
-}
-
-.icn-success {
-  color: rgb(27, 176, 27);
-}
-
-.icn-danger {
-  color: rgb(191, 61, 31);
-}
-
-.icn-disabled {
-  color: rgb(138, 138, 138);
-}
-
-.cicon {
-  font-size: 20px;
-  height: 20px;
-  width: 20px;
-  line-height: 20px;
-}
-
-a {
-  color: $appblack;
-  text-decoration: underline;
-}
-
-label {
-  display: flex;
-}
-
-.lk-checkbox-label {
-  margin-top: 0px;
 }
